+++ src/containers/Login/Register.jsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import RegisterForm from '../../components/LoginForm/RegisterForm';
 
-
 const Register = () => (
   <section className="login">
     <div className="login__container">
-      <RegisterForm/>
+      <RegisterForm />
     </div>
   </section>
 );
